+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### 04/02/2021
+-   Change `/users/create` to `/users/signup`.
+-   Add `/users/signin` endpoint.
+-   Add password encryption with argon2.
+
 ### 02/02/2021
 -   Add simple websocket server.
 -   Add `Connection` struct.